Fixed the changes requested in the pull request, internationalized the navbar and the main menu, and added the first testing cases (very trivial, as there is not much functionality yet)
