+++ webapp/src/HomePage.css
@@ -1,4 +1,4 @@
-/* MainMenu.css */
+/* HomePage.css */
 .window-container {
   display: flex;
   flex-direction: column;
