+++ webapp/src/locales/EN.json
@@ -2,10 +2,25 @@
     "about.title":"About us",
     "forms.username":"Username",
     "forms.password":"Password",
+    "gameModes.basicQuiz.description": "Play a basic quiz game. You will be presented with an image, a question related to the image, and a set of possible answers. Try to choose the correct answer. You can also ask the AI chat for a hint!",
+    "gameModes.basicQuiz.name": "Normal Game",
+    "gameModes.endlessMarathon.description": "Endless mode, no question limit. Keep answering questions until you miss while you climb up in the leaderboard!",
+    "gameModes.endlessMarathon.name": "Endless marathon",
+    "gameModes.expertDomain.description": "Play a basic game, but focused on an specific topic. Show your knowledge on your favourite topic, or let the randomized spinning wheel decide for you to make the game more challenging!",
+    "gameModes.expertDomain.name": "Expert's domain",
+    "gameModes.timeAttack.description": "In this game mode, the time is the limit. Try to answer as fast as you can and as much questions as possible before the timer goes to 0!",
+    "gameModes.timeAttack.name": "Time attack",
+    "homePage.play": "Play",
+    "homePage.randomizeOFF": "Randomize Topic: OFF",
+    "homePage.randomizeON": "Randomize Topic: ON",
+    "homePage.welcome": "Welcome to WiChat game! Select a game mode to start playing:",
     "index.title":"What is Wichat",
+    "leaderboard":"Leaderboard",
     "login.title":"Welcome back!",
     "login.error":"Invalid username or password",
     "login.message":"Log in",
     "logout":"Log off",
-    "signup.message":"Sign up"
+    "menu":"Menu",
+    "signup.message":"Sign up",
+    "statistics":"Statistics"
 }
