+++ webapp/src/locales/ES.json
@@ -2,10 +2,25 @@
     "about.title":"Quienes somos",
     "forms.username":"Usuario",
     "forms.password":"Contraseña",
+    "gameModes.basicQuiz.description": "Juega una partida rápida normal. Se te presentará una imagen junto a una pregunta relacionada con la imagen y un conjunto de posibles respuestas. ¡Intenta elegir la respuesta correcta! También puedes pedir una pista al chat de IA incorporado.",
+    "gameModes.basicQuiz.name": "Partida normal",
+    "gameModes.endlessMarathon.description": "Modo sin fin, sin límite de preguntas. ¡Sigue respondiendo preguntas hasta que falles mientras subes puestos en la tabla de clasificación!",
+    "gameModes.endlessMarathon.name": "Maratón sin fin",
+    "gameModes.expertDomain.description": "Juega una partida normal, pero enfocada en un tema específico. Demuestra tus conocimientos en un tema concreto, o deja que la ruleta de selección aleatoria decida por ti para hacer el juego más desafiante.",
+    "gameModes.expertDomain.name": "Dominio del experto",
+    "gameModes.timeAttack.description": "En este modo de juego, el tiempo es el límite. ¡Intenta responder lo más rápido que puedas y tantas preguntas como sea posible antes de que el temporizador llegue a 0!",
+    "gameModes.timeAttack.name": "Ataque de tiempo",
+    "homePage.play": "Jugar",
+    "homePage.randomizeOFF": "Tema aleatorio: NO",
+    "homePage.randomizeON": "Tema aleatorio: SÍ",
+    "homePage.welcome": "¡Bienvenid@ a WiChat! Selecciona un modo de juego para empezar a jugar:",
     "index.title":"Qué es Wichat",
+    "leaderboard":"Clasificación",
     "login.title":"Bienvenid@ de vuelta!",
     "login.error":"Inicio de sesión inválido",
     "login.message":"Iniciar sesión",
     "logout":"Cerrar sesión",
-    "signup.message":"Registrarme"
+    "menu":"Menú",
+    "signup.message":"Registrarme",
+    "statistics":"Estadísticas"
 }
